testRandomLinkNavigatesToCorrespondingPage() relocated to YPage due to Irina's request
