+++ src/test/java/tests/JTest.java
@@ -74,26 +74,5 @@ public class JTest extends BaseTest {
         Assert.assertEquals(actualUrl, expectedURL);
         Assert.assertEquals(actualTitle, expectedTitle);
     }
-
-    @Test
-    public void testRandomLinkNavigatesToCorrespondingPage() {
-
-        List<WebElement> languageNames = openBaseURL()
-                .clickBrowseLanguagesMenu()
-                .clickJSubmenu()
-                .getLanguagesLinks();
-
-        Assert.assertTrue(languageNames.size() > 0, "listSize <= 0");
-
-        int r = new Random().nextInt(languageNames.size());
-
-        JPage jPage = new JPage(getDriver());
-
-        String languageName = jPage.getTextFromRandomLink(r, languageNames).toLowerCase();
-
-        jPage.clickRandomLink(r, languageNames);
-
-        Assert.assertTrue(getExternalPageURL().contains("language-" + languageName + "-"), getExternalPageTitle() + "doesn't contain" + languageName);
-    }
 }
